requirements.txt
pandas
lyricsgenius
python-dotenv